eslint-plugin: Rename "graphql-require-description" to "graphql-require-object-description"
