+++ src/eslint-plugin-adeira/src/rules/graphql-require-object-description.js
